Use Received header rather than Date to get the received date.
When the Delivery-date header doesn't exist Maildirarc previously used
the Date header. The problem with this is that it is generated by the
sender and may be invalid or intentionally bogus. So instead it now
uses the first (ie. latest) Received header.

This required a change in the logic since Received headers are usually
split across multiple lines. The header parsing now copes with this.

Also fix some minor issues and properly document the logic used to
create the "From " separator line.