Make use of the Date header as a last resort.
Using the Date header to get the delivery date is probably a bad idea. It
could be intentionally bogus, or just badly formatted. But when looking
through messages in a sent items folder there's no Delivery-date or
Received headers to use, so we have to use it. In all other cases it
should be rare to use it; I haven't seen any emails outside of my sent
items without a Received header yet.