Add flag to skip unparseable messages.
Unparseable messages still can't be added to an mbox file. This flag
adds the option to skip them rather than fail. At the end of the
process one can now review the remaining messages and decide what to
do with them.

This mostly covers issue #1. It could be extended to allow immediate
removal of unparseable messages.